+++ changelog.md
@@ -13,6 +13,7 @@
 * Removed unnecessary event un-bindings that caused errors on Edge @rafa8626
 * Added missing layer to display errors correctly on native renderers and added missing styles for errors; `customError` accepts a callback as well @rafa8626
 * Added missing translations; removed Brazilian Portuguese and added Catalan in demo file @rafa8626
+* Added `forceLive` configuration to hide progress bar and display `Live Broadcast` even when `duration` is a valid number @rafa8626
 
 *4.2.3 (2017/07/22)*